requirements.txt
agate==1.6.3
aiodns==3.0.0
aiohttp==3.7.0
aiohttp-socks==0.4.1
alembic==1.8.0
anyio==3.6.1
apache-airflow==2.3.2
apache-airflow-providers-ftp==2.1.2
apache-airflow-providers-http==2.1.2
apache-airflow-providers-imap==2.2.3
apache-airflow-providers-sqlite==2.1.3
apispec==3.3.2
argcomplete==2.0.0
asn1crypto==1.5.1
async-timeout==3.0.1
attrs==20.3.0
Babel==2.10.1
beautifulsoup4==4.11.1
blinker==1.4
Brotli==1.0.9
bs4==0.0.1
cachelib==0.7.0
cattrs==1.10.0
cchardet==2.1.7
certifi==2020.12.5
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.3
clickclick==20.10.2
colorama==0.4.4
colorlog==4.8.0
commonmark==0.9.1
connexion==2.13.1
cron-descriptor==1.2.24
croniter==1.3.5
cryptography==3.4.8
dash==2.5.1
dash-bootstrap-components==1.1.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dataclasses==0.6
DateTime==4.4
dbt-core==1.1.0
dbt-extractor==0.4.1
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.5.1
diskcache==5.4.0
dnspython==2.2.1
docutils==0.18.1
elastic-transport==8.1.2
elasticsearch==8.2.2
email-validator==1.2.1
fake-useragent==0.1.11
filelock==3.7.1
findspark==2.0.1
finvizfinance==0.14.3
Flask==1.1.2
Flask-AppBuilder==3.4.5
Flask-Babel==2.0.0
Flask-Caching==1.11.1
Flask-Compress==1.12
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.3.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
future==0.18.2
geographiclib==1.52
geopy==2.2.0
googletransx==2.4.2
graphviz==0.20
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
hologram==0.0.14
httpcore==0.15.0
httpx==0.23.0
huggingface-hub==0.7.0
idna==3.3
importlib-metadata==4.11.4
inflection==0.5.1
isodate==0.6.1
itsdangerous==1.1.0
jeepney==0.8.0
Jinja2==2.11.3
jsonschema==3.1.1
kafka-python==2.0.2
keyring==23.6.0
lazy-object-proxy==1.7.1
leather==0.3.4
lockfile==0.12.2
Logbook==1.5.3
lxml==4.9.0
Mako==1.2.0
Markdown==3.3.7
MarkupSafe==2.0.1
marshmallow==3.16.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
mashumaro==2.9
minimal-snowplow-tracker==0.0.2
msgpack==1.0.4
multidict==6.0.2
multiprocess==0.70.13
mysqlclient==2.1.1
networkx==2.8.4
numpy==1.22.4
oscrypto==1.3.0
packaging==21.3
pandas==1.4.2
parsedatetime==2.4
pathspec==0.9.0
pendulum==2.1.2
Pillow==9.1.1
plotly==5.8.2
pluggy==1.0.0
prison==0.2.1
psutil==5.9.1
py4j==0.10.9.3
pycares==4.1.2
pycparser==2.21
pycryptodomex==3.14.1
pydruid==0.6.3
Pygments==2.12.0
PyJWT==1.7.1
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
PySocks==1.7.1
python-daemon==2.3.0
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==6.1.2
python3-openid==3.2.0
pytimeparse==1.1.8
pytz==2022.1
pytzdata==2020.1
PyYAML==5.4.1
regex==2022.6.2
requests==2.27.1
rfc3986==1.5.0
rich==12.4.4
schedule==1.1.0
SecretStorage==3.3.2
setproctitle==1.2.3
six==1.16.0
sniffio==1.2.0
snowflake-connector-python==2.7.8
soupsieve==2.3.2.post1
SQLAlchemy==1.4.9
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.38.2
sqlparse==0.4.2
swagger-ui-bundle==0.0.9
tabulate==0.8.9
tenacity==8.0.1
termcolor==1.1.0
text-unidecode==1.3
tokenizers==0.12.1
torch==1.11.0+cpu
torchaudio==0.11.0+cpu
torchvision==0.12.0+cpu
tqdm==4.64.0
transformers==4.19.4
tweet-preprocessor==0.6.0
twint==2.1.21
typing_extensions==4.2.0
unicodecsv==0.14.1
urllib3==1.26.9
Werkzeug==1.0.1
wrapt==1.14.1
WTForms==2.3.3
yarl==1.7.2
zipp==3.8.0
zope.interface==5.4.0
